Refactor of the configuration file processing and output directory handling
commit f17bd9efc7c4c099e373872b7161ad4f3769f01e
Author: Andrew <makstermaki@gmail.com>
Date:   Thu Feb 13 20:44:07 2020 -0500

    Refactor how parameters are handled
    Updated the config file so that only 2 directories able to be manually set: working directory and stream directory
      - All files needed to run the application will now be stored in subdirectories under the working directory

+++ app.py
@@ -4,8 +4,8 @@ import configparser
 import datetime
 import glob
 import os
-import re
 import subprocess
+import sys
 import time
 from pprint import pprint
 
@@ -17,6 +17,15 @@ import common.xmltv as xmltv
 
 # Throw an exception if this script is already running
 me = singleton.SingleInstance()
+
+# Default Parameters
+default_working_dir = '/tmp/HomeBroadcaster/'
+default_stream_dir = '/var/www/html/tv'
+
+# Declare the subdirectories to be used
+logs_subdir = 'logs/'
+playlist_subdir = 'playlists/'
+pid_subdir = 'pid/'
 
 db_utils.initialize_db()
 
@@ -30,6 +39,12 @@ def check_pid_running(pid):
         return True
 
 
+def clean_directory_path(path):
+    if not path.endswith('/'):
+        return path + '/'
+    return path
+
+
 def clear_previous_stream_files(channel_name, stream_dir, playlist_dir):
     file_list = glob.glob(stream_dir + channel_name + '*')
     for file in file_list:
@@ -37,11 +52,8 @@ def clear_previous_stream_files(channel_name, stream_dir, playlist_dir):
             os.remove(file)
         except:
             print("Error occurred while deleting file: ", file)
-    try:
+    if os.path.exists(playlist_dir + channel_name + ".txt"):
         os.remove(playlist_dir + channel_name + ".txt")
-    except:
-        print("Error occurred while deleting FFMPEG concat playlist: " + playlist_dir + channel_name + ".txt")
-
 
 # Given a show name, this function will populate the SQLite DB with all of the episode information
 # for that show
@@ -80,7 +92,7 @@ def populate_all_episode_info(directory_full_path):
         db_utils.update_series_last_updated_time(directory_name)
 
 
-def start_channel(channel_name, order, channel_series_id, playlist_dir, stream_dir, xmltv_file, ffmpeg_log):
+def start_channel(channel_name, order, channel_series_id, xmltv_file, dirs):
     channel_result = db_utils.get_channel(channel_name)
     if channel_result is None:
         db_utils.save_channel(channel_name, order, channel_series_id)
@@ -136,47 +148,100 @@ def start_channel(channel_name, order, channel_series_id, playlist_dir, stream_d
     db_utils.update_channel_next_episode(channel_name, last_episode_in_playlist + 1)
 
     # Generate the FFMPEG concat playlist
-    concat_playlist = playlist_utils.generate_concat_playlist(playlist, playlist_dir, channel_name)
+    concat_playlist = playlist_utils.generate_concat_playlist(playlist, dirs['playlist_dir'], channel_name)
 
     # At this point, the FFMPEG playlist has been generated so the stream can be started
     if not os.path.exists("./pid"):
         os.mkdir("./pid")
 
+    ffmpeg_log_file = open(dirs['log_dir'] + 'ffmpeg.log', "w")
+    m3u8_path = dirs['stream_dir'] + channel_name + '.m3u8'
+
     proc = subprocess.Popen([
         "ffmpeg", "-re", "-loglevel", "warning", "-fflags", "+genpts", "-f", "concat", "-safe", "0", "-i",
         concat_playlist, "-map", "0:a?", "-map", "0:v?", "-strict", "-2", "-dn", "-c", "copy",
-        "-hls_time", "10", "-hls_list_size", "6", "-hls_wrap", "7", stream_dir + channel_name + ".m3u8"
-    ], stderr=ffmpeg_log)
+        "-hls_time", "10", "-hls_list_size", "6", "-hls_wrap", "7", m3u8_path
+    ], stderr=ffmpeg_log_file)
 
-    pid_file = open("./pid/" + channel_name + ".pid", "w")
+    pid_file = open(dirs['pid_dir'] + channel_name + ".pid", "w")
     pid_file.write(str(proc.pid))
     pid_file.close()
 
     # Add the channel to the central streams playlist
-    m3u.add_channel_if_not_exists(stream_dir, channel_name)
+    m3u.add_channel_if_not_exists(dirs['stream_dir'], channel_name)
 
+
+# -----------------SCRIPT STARTS HERE---------------------
+
+try:
+    config_path = sys.argv[1]
+except Exception:
+    raise Exception('No config file passed in arguments!')
 
 config = configparser.ConfigParser()
-config.read('config.ini')
+config.read(config_path)
 
-# Retrieve already existing XML TV file or generate a new one
-xmltv_path = config.get('GENERAL', 'xmltv_path')
-if os.path.exists(xmltv_path):
-    file_xmltv = xmltv.open_xmltv(xmltv_path)
+# Check the config file for any missing parameters and generate the default directories if so
+if config.has_option('GENERAL', 'working_directory'):
+    working_directory = clean_directory_path(config.get('GENERAL', 'working_directory'))
 else:
-    file_xmltv = xmltv.generate_new_xmltv()
+    working_directory = default_working_dir
+    if not os.path.exists(working_directory):
+        os.mkdir(working_directory)
 
-playlist_directory = config.get('GENERAL', 'playlist_directory')
-stream_directory = config.get('GENERAL', 'stream_directory')
-ffmpeg_log_path = config.get('GENERAL', 'ffmpeg_log')
-ffmpeg_log_file = open(ffmpeg_log_path, "w")
+log_directory = working_directory + logs_subdir
+playlist_directory = working_directory + playlist_subdir
+pid_directory = working_directory + pid_subdir
+
+if not os.path.exists(log_directory):
+    os.mkdir(log_directory)
+if not os.path.exists(playlist_directory):
+    os.mkdir(playlist_directory)
+if not os.path.exists(pid_directory):
+    os.mkdir(pid_directory)
+
+# Populate stream directory and XML TV location
+if config.has_option('GENERAL', 'stream_directory'):
+    stream_directory = clean_directory_path(config.get('GENERAL', 'stream_directory'))
+
+    if not os.path.exists(stream_directory):
+        os.mkdir(stream_directory)
+
+    xmltv_path = stream_directory + 'xmltv.xml'
+
+    if os.path.exists(xmltv_path):
+        file_xmltv = xmltv.open_xmltv(xmltv_path)
+    else:
+        file_xmltv = xmltv.generate_new_xmltv()
+else:
+    if not os.path.exists(default_stream_dir):
+        os.mkdir(default_stream_dir)
+    stream_directory = default_stream_dir
+
+    xmltv_path = default_stream_dir + 'xmltv.xml'
+    if os.path.exists(xmltv_path):
+        file_xmltv = xmltv.open_xmltv(xmltv_path)
+    else:
+        file_xmltv = xmltv.generate_new_xmltv()
+
+# Create a dictionary for all of the working directories
+directories = {
+    'working_dir': working_directory,
+    'stream_dir': stream_directory,
+    'playlist_dir': playlist_directory,
+    'pid_dir': pid_directory,
+    'log_dir': log_directory
+}
+
+
+# All input parameters have now been processed and the channels can now be started
 
 for channel in config.sections():
     if channel == 'GENERAL':
         continue
 
     # Check if the channel is currently running and if so, do nothing
-    pid_file_path = './pid/' + channel + '.pid'
+    pid_file_path = directories['pid_dir'] + channel + '.pid'
     if os.path.exists(pid_file_path):
         old_pid_file = open(pid_file_path)
         ffmpeg_pid = old_pid_file.readline().strip()
@@ -187,7 +252,7 @@ for channel in config.sections():
         else:
             # Channel stopped running. Clear the old stream files
             print("Deleting old stream files...")
-            clear_previous_stream_files(channel, stream_directory, playlist_directory)
+            clear_previous_stream_files(channel, directories['stream_dir'], directories['playlist_dir'])
 
     print("Starting channel: " + channel)
 
@@ -197,8 +262,7 @@ for channel in config.sections():
     directory_series_id = db_utils.get_series_id(os.path.basename(directory))
 
     populate_all_episode_info(directory)
-    start_channel(channel, playback_order, directory_series_id, playlist_directory,
-                  stream_directory, file_xmltv, ffmpeg_log_file)
+    start_channel(channel, playback_order, directory_series_id, file_xmltv, directories)
 
 xmltv.remove_past_programmes(file_xmltv)
 xmltv.save_to_file(file_xmltv, xmltv_path)
